Fish

+++ DiamondShopSystem/src/main/java/com/example/DiamondShopSystem/service/JewelryService.java
@@ -44,12 +44,24 @@ public class JewelryService {
     @Autowired
     private ProductRepository productRepository;
 
-    public AllDataDTO getAllData() {
-        List<Category> categories = categoryRepository.findAll();
-        List<Material> materials = materialRepository.findAll();
-        List<Size> sizes = sizeRepository.findAll();
-        List<Shape> shapes = shapeRepository.findAll();
-        return new AllDataDTO(categories, materials, sizes, shapes);
+    @Autowired
+    private JWTUtils jwtUtils;
+    @Autowired
+    private StaffRepository staffRepository;
+
+    public AllDataDTO getAllData(String token) {
+        String username = jwtUtils.extractUsername(token);
+        Optional<Staff> staff = staffRepository.findByUsername(username);
+        if (staff == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid token");
+        } else {
+            List<Category> categories = categoryRepository.findAll();
+            List<Material> materials = materialRepository.findAll();
+            List<Size> sizes = sizeRepository.findAll();
+            List<Shape> shapes = shapeRepository.findAll();
+            return new AllDataDTO(categories, materials, sizes, shapes);
+        }
+
     }
 
     public List<Jewelry> findAllJewelry() {
@@ -65,7 +77,12 @@ public class JewelryService {
         return jewelryRepository.existsByName(name);
     }
 
-    public Jewelry saveJewelry(Jewelry jewelry) {
+    public Jewelry saveJewelry(Jewelry jewelry, String token) {
+        String username = jwtUtils.extractUsername(token);
+        Staff staff = staffRepository.findByUsernameAndRoleRoleId(username,4L);
+        if(staff == null){
+            throw new RuntimeException("this token is invalid");
+        }
         if (jewelry.getDiamond() != null) {
             Optional<Diamond> diamond = diamondRepository.findById(jewelry.getDiamond().getDiamondId());
             diamond.ifPresent(jewelry::setDiamond);
@@ -75,7 +92,12 @@ public class JewelryService {
         return savedJewelry;
     }
 
-    public Jewelry updateJewelry(Long id, Jewelry newJewelry) {
+    public Jewelry updateJewelry(Long id, Jewelry newJewelry, String token) {
+        String username = jwtUtils.extractUsername(token);
+        Staff staff = staffRepository.findByUsernameAndRoleRoleId(username,4L);
+        if(staff == null){
+            throw new RuntimeException("this token is invalid");
+        }
         return jewelryRepository.findById(id)
                 .map(existingJewelry -> {
                     existingJewelry.setName(newJewelry.getName());
@@ -90,7 +112,12 @@ public class JewelryService {
                 }).orElseThrow(() -> new RuntimeException("Jewelry not found"));
     }
 
-    public void deleteJewelry(Long id) {
+    public void deleteJewelry(Long id, String token) {
+        String username = jwtUtils.extractUsername(token);
+        Staff staff = staffRepository.findByUsernameAndRoleRoleId(username,4L);
+        if(staff == null){
+            throw new RuntimeException("this token is invalid");
+        }
         Optional<Jewelry> jewelry = jewelryRepository.findById(id);
         if (jewelry.isPresent()) {
             productService.deleteProductByJewelryId(id);
@@ -99,6 +126,7 @@ public class JewelryService {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Jewelry not found with id " + id);
         }
     }
+
     public List<Jewelry> findAllByCategoryId(Long categoryId) {
         return jewelryRepository.findAll().stream()
                 .filter(jewelry -> jewelry.getCategory().getCategoryId().equals(categoryId))
